Committed 2022/10/17

+++ Supplementary materials/TensorFlow/Back_propagate.py
@@ -5,7 +5,7 @@ import matplotlib.pyplot as plt
 np.random.seed(0)
 
 # Sample random numbers from a normal distribution with mean 1 and standard deviation of 0.1.
-x_values = np.random.normal(1, 0.1, 100).astype(np.float)
+x_values = np.random.normal(1, 0.1, 100).astype(np.float32)
 
 print("x_values: ", x_values)
